test: update unit tests
